mt-tools/pyproject.toml
[project]
name = "mt"
version = "0.1.0"
description = "Machine translation tools"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "openai-harmony",
    "unbabel-comet",
    "datasets",
    "pandas",
    "rich",
    "duckdb",
    "bert-score",
]

[project.scripts]
mt = "mt.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mt"]

